test(P10-F07): cover Active/Historic breakdown services, facade dispatch, and scope-historic sizing
Adapts the pre-split breakdown tests into ActiveBrokerBreakdownServiceTests,
adds HistoricBrokerBreakdownServiceTests covering gross-TotalBought sizing
(including the fully-closed-position case), adds BrokerBreakdownServiceTests
for facade dispatch, and extends the scope=historic integration test with
value assertions plus an Active-scope regression test.
